Ensure frame generation completion with glFinish
Since glFlush doesn't ensures all the submitted gl commands are
processed for rendering the current frame, also since eglSwapBuffers
is not used glFinish is required which ensures all the submitted
gl commands are processed for rendering the current frame. We are
using glFlush in Tizen m34 port but we didn't see this issue for
static pages as in Tizen m34 port swapbuffers call is made twice
so glFlush is working in m34, in m40 we get only one swapbuffers
call for static pages.

Original Author: m.nataraj
Original Patch: http://165.213.202.130:8080/#/c/72671/

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=9098

Change-Id: I6994c5f873d5f2f4fff9ea89d45b9056cc28db8a
Signed-off-by: m.nataraj <m.nataraj@samsung.com>
Signed-off-by: Shanmuga Pandi M <shanmuga.m@samsung.com>